fix(apps/console): resource sync updates
